Extract setRequiredErrors helper fn
