+++ src/components/ComplainantInfoView.js
@@ -11,16 +11,16 @@ import ReactRouterPropTypes from 'react-router-prop-types';
 import FormNav from './FormNav';
 import { PaddedRow, TitleLabel, SectionHeader, ErrorMessage } from '../shared/Layout';
 import { FORM_PATHS, FORM_ERRORS } from '../shared/Consts';
-import { setDidClickNav } from '../shared/Helpers';
+import { setDidClickNav, setRequiredErrors } from '../shared/Helpers';
 import { bootstrapValidation, validateRequiredInput } from '../shared/Validation';
 
-const REQUIRED_FIELDS = ['complainantName'];
 
 class ComplainantInfoView extends React.Component {
   constructor(props) {
     super(props);
 
     this.state = {
+      requiredFields: ['complainantName'],
       sectionFormatErrors: [],
       sectionRequiredErrors: [FORM_ERRORS.IS_REQUIRED],
       complainantNameValid: true,
@@ -48,36 +48,13 @@ class ComplainantInfoView extends React.Component {
     }).isRequired
   }
 
-  setRequiredErrors = () => {
-    const requiredErrors = this.state.sectionRequiredErrors.slice();
-    const areRequiredInputsValid = validateRequiredInput(
-      this.props.input,
-      REQUIRED_FIELDS
-    );
-
-    if (areRequiredInputsValid) {
-      if (requiredErrors.indexOf(FORM_ERRORS.IS_REQUIRED) !== -1) {
-        requiredErrors.splice(requiredErrors.indexOf(FORM_ERRORS.IS_REQUIRED));
-      }
-    }
-    else if (requiredErrors.indexOf(FORM_ERRORS.IS_REQUIRED) === -1) {
-      requiredErrors.push(FORM_ERRORS.IS_REQUIRED);
-    }
-
-    this.setState({
-      sectionRequiredErrors: requiredErrors
-    });
-  }
-
   handlePageChange = (path) => {
-    this.setstate(setDidClickNav);
-
-    Promise.resolve(this.setRequiredErrors())
-      .then(() => {
-        if (this.state.sectionRequiredErrors.length < 1 && this.state.sectionFormatErrors.length < 1) {
-          this.props.handlePageChange(path);
-        }
-      });
+    this.setState(setDidClickNav);
+    this.setState(setRequiredErrors, () => {
+      if (this.state.sectionRequiredErrors.length < 1 && this.state.sectionFormatErrors.length < 1) {
+        this.props.handlePageChange(path);
+      }
+    });
   }
 
   setInputErrors = (name, inputValid, sectionFormatErrors) => {
@@ -109,7 +86,7 @@ class ComplainantInfoView extends React.Component {
   componentWillUnmount() {
     const areRequiredInputsValid = validateRequiredInput(
       this.props.input,
-      REQUIRED_FIELDS
+      this.state.requiredFields
     );
     if (
       !areRequiredInputsValid
