+++ src/components/DispositionView.js
@@ -19,10 +19,8 @@ import {
   ErrorMessage
 } from '../shared/Layout';
 import { FORM_PATHS, FORM_ERRORS } from '../shared/Consts';
-import { setDidClickNav } from '../shared/Helpers';
+import { setDidClickNav, setRequiredErrors } from '../shared/Helpers';
 import { bootstrapValidation, validateRequiredInput } from '../shared/Validation';
-
-const REQUIRED_FIELDS = ['disposition', 'incidentNarrative'];
 
 
 class DispositionView extends React.Component {
@@ -30,6 +28,7 @@ class DispositionView extends React.Component {
     super(props);
 
     this.state = {
+      requiredFields: ['disposition', 'incidentNarrative'],
       sectionFormatErrors: [],
       sectionRequiredErrors: [FORM_ERRORS.IS_REQUIRED],
       dispositionValid: true,
@@ -63,36 +62,13 @@ class DispositionView extends React.Component {
     }).isRequired
   }
 
-  setRequiredErrors = () => {
-    const requiredErrors = this.state.sectionRequiredErrors.slice();
-    const areRequiredInputsValid = validateRequiredInput(
-      this.props.input,
-      REQUIRED_FIELDS
-    );
-
-    if (areRequiredInputsValid) {
-      if (requiredErrors.indexOf(FORM_ERRORS.IS_REQUIRED) !== -1) {
-        requiredErrors.splice(requiredErrors.indexOf(FORM_ERRORS.IS_REQUIRED));
-      }
-    }
-    else if (requiredErrors.indexOf(FORM_ERRORS.IS_REQUIRED) === -1) {
-      requiredErrors.push(FORM_ERRORS.IS_REQUIRED);
-    }
-
-    this.setState({
-      sectionRequiredErrors: requiredErrors
-    });
-  }
-
   handlePageChange = (path) => {
     this.setState(setDidClickNav);
-
-    Promise.resolve(this.setRequiredErrors())
-      .then(() => {
-        if (this.state.sectionRequiredErrors.length < 1 && this.state.sectionFormatErrors.length < 1) {
-          this.props.handlePageChange(path);
-        }
-      });
+    this.setState(setRequiredErrors, () => {
+      if (this.state.sectionRequiredErrors.length < 1 && this.state.sectionFormatErrors.length < 1) {
+        this.props.handlePageChange(path);
+      }
+    });
   }
 
   setInputErrors = (name, inputValid, sectionFormatErrors) => {
@@ -125,7 +101,7 @@ class DispositionView extends React.Component {
   componentWillUnmount() {
     const areRequiredInputsValid = validateRequiredInput(
       this.props.input,
-      REQUIRED_FIELDS
+      this.state.requiredFields
     );
     if (
       !areRequiredInputsValid
