+++ src/shared/Helpers.js
@@ -1,5 +1,29 @@
+import { bootstrapValidation, validateRequiredInput } from '../shared/Validation';
+import { FORM_PATHS, FORM_ERRORS } from '../shared/Consts';
+
 export const setDidClickNav = () => {
   return {
     didClickNav: true
   };
 };
+
+export const setRequiredErrors = (state, props) => {
+  const requiredErrors = state.sectionRequiredErrors.slice();
+  const areRequiredInputsValid = validateRequiredInput(
+    props.input,
+    state.requiredFields
+  );
+
+  if (areRequiredInputsValid) {
+    if (requiredErrors.indexOf(FORM_ERRORS.IS_REQUIRED) !== -1) {
+      requiredErrors.splice(requiredErrors.indexOf(FORM_ERRORS.IS_REQUIRED));
+    }
+  }
+  else if (requiredErrors.indexOf(FORM_ERRORS.IS_REQUIRED) === -1) {
+    requiredErrors.push(FORM_ERRORS.IS_REQUIRED);
+  }
+
+  return {
+    sectionRequiredErrors: requiredErrors
+  };
+}
